balance tweaks to most levels, changes to how spikes work
Spikes are now done in two layers - the actual Death layer, and a "cosmetic" layer which uses the larger spikes. I also put in a change for the sprites we use for the spikes that generate the collider (it's smaller and should be less prone to accidentally run into from below). Platforms still kind of fucked. Probably need some solution collider.